Updated scatter plots
All names are in the following format:
scatter_ (1. all/gb) _ (2. orig/ch_no/4ch/idx) _ 3. (avg/max)

1. All => shows all classes. gb => shows only good/bad
2. orig => original spectrogram. ch_no is if it is for a particular channel. 4ch is an operation on all 4 channels. idx is if it is an operation on the acoustic indices.
3. avg and max.

Eg: Scatter_gb_4ch_avg is a scatterplot showing the good and bad classes, where the 4 channels have been averaged and the indices computed on that average.
Scatter_all_idx_max would be a scatterplot where all 4 classes are shown. The indices are computed for the 4 channels and max is taken of the indices.